Hook up small components for Team Members tab & move Metrics Reporting Agency component
