+++ publisher/src/components/AdminPanel/AgencyProvisioningComponents/TeamMembersSearch.tsx
@@ -1,0 +1,81 @@
+// Recidiviz - a data platform for criminal justice reform
+// Copyright (C) 2024 Recidiviz, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+// =============================================================================
+
+import { observer } from "mobx-react-lite";
+import React, { useEffect, useState } from "react";
+
+import AdminPanelStore from "../../../stores/AdminPanelStore";
+import * as Styled from "../AdminPanel.styles";
+import { useAgencyProvisioning } from "../AgencyProvisioningContext";
+import { AgencyTeamMembers } from "../types";
+import { getCurrentTeamMembers } from "./utils";
+
+type TeamMembersSearchType = {
+  onFilteredChange: (filtered: AgencyTeamMembers) => void;
+};
+
+export const TeamMembersSearch: React.FC<TeamMembersSearchType> = observer(
+  ({ onFilteredChange }) => {
+    const { selectedAgency } = useAgencyProvisioning();
+
+    const [teamMembersSearchInput, setTeamMembersSearchInput] =
+      useState<string>("");
+
+    const currentTeamMembers = getCurrentTeamMembers(selectedAgency);
+    const filteredTeamMembers = AdminPanelStore.searchList(
+      currentTeamMembers,
+      teamMembersSearchInput,
+      ["id", "name", "email"]
+    );
+
+    useEffect(() => {
+      onFilteredChange(filteredTeamMembers);
+    }, [filteredTeamMembers, onFilteredChange]);
+
+    return (
+      <Styled.InputLabelWrapper noBottomSpacing>
+        <input
+          id="search-team-members"
+          name="search-team-members"
+          type="text"
+          value={teamMembersSearchInput}
+          onChange={(e) => setTeamMembersSearchInput(e.target.value)}
+        />
+        <Styled.LabelWrapper>
+          <label htmlFor="search-team-members">
+            Search by name, email or user ID
+          </label>
+          {teamMembersSearchInput && (
+            <Styled.LabelButton
+              onClick={() => {
+                setTeamMembersSearchInput("");
+              }}
+            >
+              Clear
+            </Styled.LabelButton>
+          )}
+        </Styled.LabelWrapper>
+        {!filteredTeamMembers.length && (
+          <Styled.EmptySearchMessage>
+            No current team members found. Please modify your search and try
+            again.
+          </Styled.EmptySearchMessage>
+        )}
+      </Styled.InputLabelWrapper>
+    );
+  }
+);
