+++ publisher/src/components/AdminPanel/index.ts
@@ -23,4 +23,3 @@ export * from "./InteractiveSearchList";
 export * from "./AdminPanel";
 export * from "./types";
 export * from "./SaveConfirmation";
-export * from "./MetricsReportingAgency";
